requirements.txt
streamlit
torch
torchvision
torchaudio
opencv-python
numpy==1.26.4
pillow
matplotlib
